refactor: move sosDraft in correct folder
